+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - Add sample configuration file to deb and txz packages
 - Debian script for cross compilation environments
+- Don't remove local backup with less than 24 hours
 
 ### Changed
 - Archive algorithm refactory to simplify the tar file